requirements.txt
PyYAML == 5.3
docopt == 0.6.2
bottle == 0.12.18
